Changed test infrastructure.
